move BAEL-1565 code to testing-modules/mocks/mock-comparisons
