+++ testing-modules/mocks/mock-comparisons/pom.xml
@@ -14,7 +14,7 @@
 
     <properties>
         <mockito.version>2.9.0</mockito.version>
-        <easymock.version>3.4</easymock.version>
+        <easymock.version>3.5.1</easymock.version>
         <jmockit.version>1.34</jmockit.version>
 
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
