+++ testing-modules/mocks/mock-comparisons/src/main/java/com/baeldung/easymock/ArticleReader.java
@@ -1,4 +1,4 @@
-package com.baeldung;
+package com.baeldung.easymock;
 
 import java.util.ArrayList;
 import java.util.Iterator;
