+++ testing-modules/mocks/mock-comparisons/src/main/java/com/baeldung/easymock/BaeldungArticle.java
@@ -1,4 +1,4 @@
-package com.baeldung;
+package com.baeldung.easymock;
 
 public class BaeldungArticle {
