+++ testing-modules/mocks/mock-comparisons/src/main/java/com/baeldung/easymock/BaeldungReader.java
@@ -1,4 +1,4 @@
-package com.baeldung;
+package com.baeldung.easymock;
 
 import java.util.List;
