+++ testing-modules/mocks/mock-comparisons/src/test/java/com/baeldung/easymock/BaeldungReaderMockDelegationTest.java
@@ -1,4 +1,4 @@
-package com.baeldung;
+package com.baeldung.easymock;
 
 import org.easymock.*;
 import org.junit.*;
